+++ src/App.tsx
@@ -1,0 +1,21 @@
+﻿import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { HomePage } from './pages/HomePage';
+import { AppRoutes } from './Constants';
+import { OfferPage } from './pages/OfferPage';
+import { LoginPage } from './pages/LoginPage';
+import { FavouritesPage } from './pages/FavouritesPage';
+import Layout from './components/Layout';
+
+export function App() {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route element={<Layout />}>
+          <Route index path={AppRoutes.Main} element={<HomePage cardsCount={5} />} />
+          <Route path={AppRoutes.Offer} element={<OfferPage />} />
+          <Route path={AppRoutes.Favorites} element={<FavouritesPage />} />
+          <Route path={AppRoutes.Login} element={<LoginPage />} />
+        </Route>
+      </Routes>
+    </BrowserRouter>);
+}
