+++ src/Constants.ts
@@ -1,0 +1,6 @@
+﻿export const AppRoutes = {
+  Main: '/',
+  Login: '/login',
+  Offer: '/offer/:id',
+  Favorites: '/favorites'
+};
