--- src/components/App.tsx
@@ -1,5 +1,0 @@
-﻿import { HomePage } from './HomePage';
-
-export function App() {
-  return <HomePage cardsCount={5} />;
-}
